requirements.txt
amqp==5.0.2
asgiref==3.3.1
autopep8==1.5.4
beautifulsoup4==4.9.3
billiard==3.6.3.0
blis==0.7.4
catalogue==1.0.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cymem==2.0.5
Django==3.1.4
django-celery-beat==2.1.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
en-core-web-sm==2.3.1
idna==2.10
importlib-metadata==3.3.0
joblib==1.0.0
kombu==5.0.2
murmurhash==1.0.5
nltk==3.5
numpy==1.19.4
plac==1.1.3
preshed==3.0.5
prompt-toolkit==3.0.8
pyahocorasick==1.4.0
pycodestyle==2.6.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
redis==3.5.3
regex==2020.11.13
requests==2.25.1
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
soupsieve==2.1
spacy==2.3.5
spacy-lookups-data==0.3.2
sqlparse==0.4.1
srsly==1.0.5
textsearch==0.0.17
thinc==7.4.5
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.55.0
typing-extensions==3.7.4.3
Unidecode==1.1.2
urllib3==1.26.2
vine==5.0.0
wasabi==0.8.0
wcwidth==0.2.5
zipp==3.4.0
